+++ src/routes/mod.rs
@@ -2,6 +2,7 @@ pub mod admin;
 pub mod book;
 pub mod health;
 pub mod library;
+pub mod static_route;
 
 use axum::Router;
 
@@ -13,5 +14,6 @@ pub fn router(state: AppState) -> Router {
         .merge(library::routes())
         .merge(book::routes())
         .merge(admin::routes())
+        .merge(static_route::routes())
         .with_state(state)
 }